requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asyncio==4.0.0
attrs==25.3.0
certifi==2025.8.3
dnspython==2.7.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
motor==3.7.1
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
pyrogrammod==2.2.1
PySocks==1.7.1
setuptools==69.1.1
typing-inspection==0.4.1
typing_extensions==4.14.1
wheel==0.43.0
yarl==1.20.1
